Added readme and moved things around
