+++ fud2/scripts/calyx-to-report.rhai
@@ -33,10 +33,8 @@ fn synth_setup(e) {
     e.rsrc("synth.tcl");
 
     // Python scripts for parsing reports for visualization and extracting JSON summary
-    e.var_("parse-rpt-script", "$calyx-base/tools/report-parsing/rpt.py");
-    e.rule("parse-rpt", "python3 $parse-rpt-script $in > $out");
-    e.var_("extract-rpt-script", "$calyx-base/tools/report-parsing/extract.py");
-    e.rule("extract-rpt", "python3 $extract-rpt-script > $out");
+    e.rule("parse-rpt", "synthrep viz -t flamegraph -f $in > $out");
+    e.rule("extract-rpt", "synthrep summary > $out");
 
     // Bash script to create a FlameGraph
     e.config_var("flamegraph-script", "flamegraph.script");
