+++ tools/report-parsing/pyproject.toml
@@ -3,12 +3,13 @@ requires = ["flit_core >=3.2,<4"]
 build-backend = "flit_core.buildapi"
 
 [project]
-name = "parser"
+name = "synthrep"
 authors = [{ name = "The Calyx Authors" }]
 license = { file = "LICENSE" }
 classifiers = ["License :: OSI Approved :: MIT License"]
 dynamic = ["version", "description"]
 dependencies = ["pandas", "plotly"]
+readme = "README.md"
 
 [project.scripts]
 synthrep = "parser.cli:main"
